Menu added: menu.py, arcanoid.py updated
In menu.py, a menu system has been added to start and finish game. The file includes the BaseMenu class as the main class, with MainMenu and GameOverMenu as its subclasses.

In arcanoid.py, functionality for controlling the paddle speed has been added using the up and down arrow keys. Additionally, a counter of missed balls was added. The integration of the menu system is also included.